Split functionality between HaloCatalog and AnalysisPipeline classes.
