+++ yt_astro_analysis/halo_analysis/__init__.py
@@ -1,17 +1,17 @@
 from yt_astro_analysis.halo_analysis.halo_catalog.halo_catalog import \
     HaloCatalog
 
-from yt_astro_analysis.halo_analysis.halo_catalog.halo_callbacks import \
-    add_callback
+from yt_astro_analysis.halo_analysis.halo_catalog.analysis_operators import \
+    add_quantity
 
 from yt_astro_analysis.halo_analysis.halo_catalog.halo_finding_methods import \
     add_finding_method
 
+from yt_astro_analysis.halo_analysis.halo_catalog.halo_callbacks import \
+    add_callback
+
 from yt_astro_analysis.halo_analysis.halo_catalog.halo_filters import \
     add_filter
-
-from yt_astro_analysis.halo_analysis.halo_catalog.halo_quantities import \
-    add_quantity
 
 from yt_astro_analysis.halo_analysis.halo_catalog.halo_recipes import \
     add_recipe
