+++ yt_astro_analysis/halo_analysis/halo_catalog/halo_filters.py
@@ -1,45 +1,17 @@
 """
-Halo filter object
+HaloCatalog filters
 
 
 
 """
 
-#-----------------------------------------------------------------------------
-# Copyright (c) yt Development Team. All rights reserved.
-#
-# Distributed under the terms of the Modified BSD License.
-#
-# The full license is in the file COPYING.txt, distributed with this software.
-#-----------------------------------------------------------------------------
-
 import numpy as np
 
-from yt.utilities.operator_registry import \
-    OperatorRegistry
 from yt.utilities.on_demand_imports import \
     _scipy as scipy
 
-from yt_astro_analysis.halo_analysis.halo_catalog.halo_callbacks import \
-    HaloCallback
-
-filter_registry = OperatorRegistry()
-
-def add_filter(name, function):
-    filter_registry[name] = HaloFilter(function)
-
-class HaloFilter(HaloCallback):
-    r"""
-    A HaloFilter is a function that minimally takes a Halo object, performs 
-    some analysis, and returns either True or False.  The return value determines 
-    whether the Halo is added to the final halo catalog being generated by the 
-    HaloCatalog object.
-    """
-    def __init__(self, function, *args, **kwargs):
-        HaloCallback.__init__(self, function, args, kwargs)
-
-    def __call__(self, halo):
-        return self.function(halo, *self.args, **self.kwargs)
+from yt_astro_analysis.halo_analysis.halo_catalog.analysis_operators import \
+    add_filter
 
 def quantity_value(halo, field, operator, value, units):
     r"""
