+++ yt_astro_analysis/halo_analysis/halo_catalog/halo_object.py
@@ -5,19 +5,16 @@ Halo object.
 
 """
 
-#-----------------------------------------------------------------------------
-# Copyright (c) 2013, yt Development Team.
-#
-# Distributed under the terms of the Modified BSD License.
-#
-# The full license is in the file COPYING.txt, distributed with this software.
-#-----------------------------------------------------------------------------
+class Halo:
+    def __init__(self, halo_catalog, data_source, index):
+        self.halo_catalog = halo_catalog
+        self.data_source = data_source
+        self.index = index
+        self.quantities = {}
 
-from yt_astro_analysis.halo_analysis.halo_catalog.analysis_pipeline import \
-    AnalysisTarget
+    def _get_field_value(self, fieldname):
+        return self.data_source[fieldname][self.index]
 
-class Halo(AnalysisTarget):
-    _container_name = "halo_catalog"
-
-    def _get_field_value(self, fieldname, data_source, index):
-        return data_source[fieldname][index]
+    def _set_field_value(self, fieldkey, fieldname):
+        self.quantities[fieldkey] = \
+          self._get_field_value(fieldname)
