+++ yt_astro_analysis/halo_analysis/halo_finding/api.py
@@ -15,7 +15,7 @@ API for halo_finding
 
 import warnings
 
-from yt.funcs import VisibleDeprecationWarning
+from numpy import VisibleDeprecationWarning
 
 warnings.warn(
     "Running the HoP and FoF halo finders directly is deprecated. " +
