+++ yt_astro_analysis/halo_analysis/halo_finding/rockstar/__init__.py
@@ -1,0 +1,2 @@
+from yt_astro_analysis.halo_analysis.halo_finding.rockstar.rockstar import \
+    RockstarHaloFinder
