+++ yt_astro_analysis/halo_analysis/tests/test_halo_catalog.py
@@ -18,7 +18,7 @@ import os
 import shutil
 import tempfile
 
-from yt_astro_analysis.halo_analysis.api import \
+from yt_astro_analysis.halo_analysis import \
     HaloCatalog, \
     add_quantity
 from yt.loaders import \
