+++ yt_astro_analysis/halo_analysis/tests/test_halo_finders.py
@@ -6,6 +6,8 @@ import tempfile
 from yt.loaders import load
 from yt.frontends.halo_catalog.data_structures import \
     YTHaloCatalogDataset
+from yt.frontends.rockstar.data_structures import \
+    RockstarDataset
 from yt.utilities.answer_testing.framework import \
     FieldValuesTest, \
     requires_ds
@@ -24,29 +26,38 @@ etiny = "enzo_tiny_cosmology/DD0046/DD0046"
 def test_halo_analysis_finders():
     from mpi4py import MPI
 
+    tmpdir = tempfile.mkdtemp()
     curdir = os.getcwd()
+    os.chdir(tmpdir)
+
     filename = os.path.join(os.path.dirname(__file__),
                             "run_halo_finder.py")
     for method in methods:
-        if method == "rockstar":
-            from nose import SkipTest
-            raise SkipTest
-        tmpdir = tempfile.mkdtemp()
-        os.chdir(tmpdir)
-        comm = MPI.COMM_SELF.Spawn(sys.executable,
-                                   args=[filename, method, tmpdir],
-                                   maxprocs=methods[method])
+        comm = MPI.COMM_SELF.Spawn(
+            sys.executable,
+            args=[filename, method, tmpdir],
+            maxprocs=methods[method])
         comm.Disconnect()
 
-        fn = os.path.join(tmpdir, "halo_catalogs", method,
-                          "%s.0.h5" % method)
+        if method == "rockstar":
+            hcfn = "halos_0.0.bin"
+        else:
+            hcfn = os.path.join("DD0046", "DD0046.0.h5")
+        fn = os.path.join(tmpdir, "halo_catalogs", method, hcfn)
+
         ds = load(fn)
         if method == "rockstar":
             ds.parameters['format_revision'] = 2
-        assert isinstance(ds, YTHaloCatalogDataset)
-        for field in _fields:
-            yield FieldValuesTest(ds, field, particle_type=True,
-                                  decimals=decimals[method])
+            ds_type = RockstarDataset
+        else:
+            ds_type = YTHaloCatalogDataset
+        assert isinstance(ds, ds_type)
 
-        os.chdir(curdir)
-        shutil.rmtree(tmpdir)
+        for field in _fields:
+            my_test = FieldValuesTest(ds, field, particle_type=True,
+                                      decimals=decimals[method])
+            my_test.suffix = method
+            yield my_test
+
+    os.chdir(curdir)
+    shutil.rmtree(tmpdir)
